Move widget tests to their appropriate package and class (fixes #794)
