+++ tests/org.jboss.reddeer.swt.test/src/org/jboss/reddeer/swt/test/shell/WorkbenchShellTest.java
@@ -1,0 +1,30 @@
+package org.jboss.reddeer.swt.test.shell;
+
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
+
+import org.jboss.reddeer.junit.runner.RedDeerSuite;
+import org.jboss.reddeer.swt.api.Shell;
+import org.jboss.reddeer.workbench.impl.shell.WorkbenchShell;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+@RunWith(RedDeerSuite.class)
+public class WorkbenchShellTest {
+	
+	@Test
+	public void workbenchShellTest() {
+		Shell shell = new WorkbenchShell();
+		assertFalse(shell.getText().equals(""));
+	}
+	
+	@Test
+	public void maximizeWorkbenshShellTest() {
+		WorkbenchShell workbenchShell = new WorkbenchShell();
+		assertFalse(workbenchShell.isMaximized());
+		workbenchShell.maximize();
+		assertTrue(workbenchShell.isMaximized());
+		workbenchShell.restore();
+		assertFalse(workbenchShell.isMaximized());
+	}
+}
