+++ tests/org.jboss.reddeer.swt.test/src/org/jboss/reddeer/swt/test/toolbar/ViewToolBarTest.java
@@ -1,0 +1,57 @@
+package org.jboss.reddeer.swt.test.toolbar;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertTrue;
+
+import org.jboss.reddeer.common.matcher.RegexMatcher;
+import org.jboss.reddeer.core.matcher.WithTooltipTextMatcher;
+import org.jboss.reddeer.junit.runner.RedDeerSuite;
+import org.jboss.reddeer.swt.api.ToolBar;
+import org.jboss.reddeer.swt.api.ToolItem;
+import org.jboss.reddeer.swt.impl.toolbar.DefaultToolItem;
+import org.jboss.reddeer.swt.impl.toolbar.ViewToolBar;
+import org.jboss.reddeer.workbench.impl.shell.WorkbenchShell;
+import org.jboss.reddeer.workbench.impl.view.WorkbenchView;
+import org.jboss.tools.reddeer.swt.test.model.TestModel;
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+@RunWith(RedDeerSuite.class)
+public class ViewToolBarTest {
+
+	@Before
+	public void prepare() {
+		new WorkbenchView("RedDeer SWT").open();
+	}
+	
+	@Test
+	public void testViewToolBar() {
+		new WorkbenchShell();
+		ToolBar t = new ViewToolBar();
+		assertNotNull(t);
+	}
+	
+	@Test
+	public void testToolItemInViewToolBarFound() {
+		ToolItem i = new DefaultToolItem("RedDeer SWT ViewToolItem");
+		assertEquals("RedDeer SWT ViewToolItem", i.getToolTipText());
+	}
+
+	@Test
+	public void testToolItemInViewToolBarClicked() {
+		ToolItem i = new DefaultToolItem("RedDeer SWT ViewToolItem");
+		i.click();		
+		assertTrue("ToolItem should be clicked", TestModel.getClickedAndReset());		
+	}
+
+	@Test
+	public void testToolItemInViewToolBarRegexClicked() {
+		WithTooltipTextMatcher rm = new WithTooltipTextMatcher(
+				new RegexMatcher("RedDeer SWT View.*"));
+		ToolItem i = new DefaultToolItem(rm);
+		i.click();
+		assertTrue("ToolItem should be clicked", TestModel.getClickedAndReset());		
+	}	
+}
